+++ src/client/src/commons/form/profile/components/profileLayers/components/profileLayersValidation/styles.js
@@ -7,7 +7,9 @@ export const LayerContainer = styled.div`
 `;
 
 export const Layer = styled.div`
-  box-shadow: inset -1px 0 0 lightgrey, inset 0 -1px 0 lightgrey;
-  flex: "1 1 0px",
-  overflowY: "auto",
+  box-shadow:
+    inset -1px 0 0 lightgrey,
+    inset 0 -1px 0 lightgrey;
+  flex: 1 1 0;
+  overflow-y: "auto";
 `;