+++ src/client/src/commons/form/workgroup/radio.jsx
@@ -7,7 +7,6 @@ const WorkgroupRadioGroup = props => {
     <Form
       size="tiny"
       style={{
-        border: "1px solid #e0e0e0",
         padding: "0.5em",
       }}>
       <Form.Field>