+++ src/client/src/commons/map/mapStyleFunctions.js
@@ -34,10 +34,20 @@ const outerSelectedStyle = new Style({
   }),
 });
 
+export const drawStyle = new Style({
+  fill: new Fill({
+    color: "rgba(255, 255, 255, 0.5)",
+  }),
+  stroke: new Stroke({
+    color: "#0099ff",
+    width: 3,
+  }),
+});
+
 export function styleFunction(feature, highlighted) {
   let selected = highlighted !== undefined && highlighted.length > 0 && highlighted.indexOf(feature.getId()) > -1;
   let res = feature.get("restriction");
-  let fill = null;
+  let fill;
   if (res === "f") {
     fill = greenFill;
   } else if (["b", "g"].indexOf(res) >= 0) {
@@ -46,7 +56,7 @@ export function styleFunction(feature, highlighted) {
     fill = blackFill;
   }
 
-  let conf = null;
+  let conf;
   let kind = feature.get("kind");
   if (kind === "B") {
     // boreholes