+++ src/client/src/commons/menu/editor/sideDrawer.tsx
@@ -2,7 +2,7 @@ import { Box, Collapse, Paper } from "@mui/material";
 import { SideDrawerProps } from "./menuComponents/menuComponentInterfaces";
 import { theme } from "../../../AppTheme";
 
-export const SideDrawer = ({ drawerOpen, drawerWidth, drawerContent }: SideDrawerProps) => {
+export const SideDrawer = ({ drawerOpen, drawerContent }: SideDrawerProps) => {
   return (
     <Box
       sx={{
@@ -14,9 +14,10 @@ export const SideDrawer = ({ drawerOpen, drawerWidth, drawerContent }: SideDrawe
       <Collapse in={drawerOpen} orientation="horizontal">
         <Paper
           sx={{
-            width: drawerWidth,
-            height: "100vh",
-            padding: 2,
+            width: "360px",
+            height: "calc(100vh - 70px)",
+            padding: "16px",
+            backgroundColor: theme.palette.background.lightgrey,
             boxShadow: theme.palette.boxShadow + " 2px 6px 6px 0px",
             borderRight: "1px solid rgba(0, 0, 0, 0.12)",
           }}>