+++ src/client/src/commons/menu/editor/sideDrawerHeader.tsx
@@ -1,0 +1,24 @@
+import { Box, IconButton, Typography } from "@mui/material";
+import { theme } from "../../../AppTheme";
+import CloseIcon from "@mui/icons-material/Close";
+
+export function SideDrawerHeader({ title, toggleDrawer }: { title: string; toggleDrawer: (open: boolean) => void }) {
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        height: "70px",
+        alignItems: "center",
+        justifyContent: "flex-start",
+        borderBottom: "1px solid " + theme.palette.border,
+        marginBottom: "24px",
+      }}>
+      <Typography variant="h5" sx={{ flexGrow: 1 }}>
+        {title}
+      </Typography>
+      <IconButton onClick={() => toggleDrawer(false)} sx={{ marginRight: "6px" }}>
+        <CloseIcon />
+      </IconButton>
+    </Box>
+  );
+}