+++ src/client/src/commons/search/components/listFilterStyles.js
@@ -18,16 +18,11 @@ export const AttributesContainer = styled(Form.Field)``;
 export const Label = styled.label`
   font-weight: 700 !important;
   font-size: 12px !important;
+  margin-top: 10px !important;
 `;
 
 export const AttributesItem = styled.div`
   width: 100%;
   margin-top: 5px;
-`;
-
-export const Reset = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: end;
-  font-size: 12px !important;
+  margin-bottom: 10px;
 `;