+++ src/client/src/commons/search/editor/FilterInterfaces.ts
@@ -1,0 +1,9 @@
+export interface FilterChipsProps {
+  activeFilters: Filter[];
+  setFilter: (key: string, value: string | boolean | number | null) => void;
+}
+
+export interface Filter {
+  key: string;
+  value: string | boolean | number | null;
+}