+++ src/client/src/pages/editor/editorComponent.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
 import { Route, Switch, useLocation, withRouter } from "react-router-dom";
 import BoreholeForm from "../../commons/form/borehole/boreholeForm";
 import MainSideNav from "../../commons/menu/editor/mainSideNav.tsx";
@@ -10,8 +10,11 @@ import { SideDrawer } from "../../commons/menu/editor/sideDrawer.tsx";
 import { Box } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { theme } from "../../AppTheme";
-import SearchEditorComponent from "../../commons/search/editor/searchEditorComponent.jsx";
+import FilterComponent from "../../commons/search/editor/filterComponent.jsx";
 import NewBoreholePanel from "../../commons/menu/editor/newBoreholePanel.tsx";
+import { DrawerContentTypes } from "./editorComponentInterfaces.ts";
+import { AlertContext } from "../../components/alert/alertContext.tsx";
+import CustomLayersPanel from "../../commons/menu/editor/customLayers/customLayersPanel.jsx";
 
 const AppBox = styled(Box)({
   display: "flex",
@@ -48,19 +51,8 @@ const EditorComponent = props => {
   const location = useLocation();
   const [workgroup, setWorkgroup] = useState(0);
   const [enabledWorkgroups, setEnabledWorkgroups] = useState([]);
-  const [sideDrawerContent, setSideDrawerContent] = useState("searchEditor");
-
-  const sideDrawerComponentMap = {
-    filters: <SearchEditorComponent />,
-    newBorehole: (
-      <NewBoreholePanel
-        workgroup={workgroup}
-        setWorkgroup={setWorkgroup}
-        enabledWorkgroups={enabledWorkgroups}
-        setEnabledWorkgroups={setEnabledWorkgroups}
-      />
-    ),
-  };
+  const [sideDrawerContent, setSideDrawerContent] = useState(DrawerContentTypes.Filters);
+  const alertContext = useContext(AlertContext);
 
   const toggleSideDrawer = open => {
     setSideDrawerOpen(open);
@@ -68,6 +60,20 @@ const EditorComponent = props => {
 
   const toggleBottomDrawer = open => {
     setBottomDrawerOpen(open);
+  };
+
+  const sideDrawerComponentMap = {
+    filters: <FilterComponent toggleDrawer={toggleSideDrawer} />,
+    newBorehole: (
+      <NewBoreholePanel
+        toggleDrawer={toggleSideDrawer}
+        workgroup={workgroup}
+        setWorkgroup={setWorkgroup}
+        enabledWorkgroups={enabledWorkgroups}
+        setEnabledWorkgroups={setEnabledWorkgroups}
+      />
+    ),
+    customLayers: <CustomLayersPanel toggleDrawer={toggleSideDrawer} />,
   };
 
   useEffect(() => {
@@ -100,11 +106,7 @@ const EditorComponent = props => {
             <Route component={DetailSideNav} path="/:id" />
           </Switch>
         </SidebarBox>
-        <SideDrawer
-          drawerOpen={sideDrawerOpen}
-          drawerWidth={240}
-          drawerContent={sideDrawerComponentMap[sideDrawerContent]}
-        />
+        <SideDrawer drawerOpen={sideDrawerOpen} drawerContent={sideDrawerComponentMap[sideDrawerContent]} />
         <MainContentBox>
           <Switch>
             <Route
@@ -117,6 +119,7 @@ const EditorComponent = props => {
                   setSort={setSort}
                   toggleBottomDrawer={toggleBottomDrawer}
                   bottomDrawerOpen={bottomDrawerOpen}
+                  displayErrorMessage={alertContext.error}
                 />
               )}
             />