+++ src/client/src/pages/editor/editorComponentInterfaces.ts
@@ -1,4 +1,5 @@
 export enum DrawerContentTypes {
   Filters = "filters",
   NewBorehole = "newBorehole",
+  CustomLayers = "customLayers",
 }