+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Changed sidebar layout and design.
 - Moved form to add a borehole from modal to sidepanel.
 - Removed fields `Inclination` `InclinationDirection` and `QtInclinationDirectionId` from borehole.
+- Removed filter by map functionality.
+- Moved icon to handle custom layers to sidebar.
+- Show more entries on the boreholes table when screensize is smaller than 1200px.
 - Calculate TVD values in General panel using the geometry of the borehole.
 
 ### Fixed